+++ hc/accounts/tests/test_senddeletionnotices.py
@@ -10,6 +10,7 @@ from django.utils.timezone import now
 from hc.accounts.management.commands.senddeletionnotices import Command
 from hc.accounts.models import Member
 from hc.api.models import Check, Ping
+from hc.payments.models import Subscription
 from hc.test import BaseTestCase
 
 
@@ -75,10 +76,9 @@ class SendDeletionNoticesTestCase(BaseTestCase):
         result = Command(stdout=Mock()).handle()
         self.assertEqual(counts(result), [0, 0, 0])
 
-    def test_it_checks_sms_limit(self):
-        # alice has a paid account
-        self.profile.sms_limit = 50
-        self.profile.save()
+    def test_it_checks_subscription(self):
+        # alice has a subscription
+        Subscription.objects.create(user=self.alice, subscription_id="abc123")
 
         result = Command(stdout=Mock()).handle()
         self.assertEqual(counts(result), [0, 0, 0])